# project-docs/openapi.yaml
openapi: 3.0.3
info:
  title: Forge API
  description: |
    REST API for Forge, a file-system-backed project management tool for software teams.
    All request/response bodies use **snake_case** field names.
    Paginated list endpoints return `{ items, total, page, limit, total_pages }`.
  version: 1.0.0
servers:
  - url: http://localhost:3001
    description: Local development server

# ─────────────────────────────────────────────────────────────────────────────
# Paths
# ─────────────────────────────────────────────────────────────────────────────
paths:
...
  /api/wiki:
    get:
      summary: List wiki docs (paginated)
      operationId: listWikiDocs
      tags: [Wiki]
      parameters:
        - $ref: "#/components/parameters/page"
        - name: limit
          in: query
          schema: { type: integer, default: 50, minimum: 1, maximum: 100 }
          description: Items per page (default 50)
      responses:
        "200":
          description: Paginated list of wiki docs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedWikiDocs"

    post:
      summary: Create a wiki doc (auto-generated slug)
      operationId: createWikiDoc
      tags: [Wiki]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveWikiDocInput"
      responses:
        "201":
          description: Doc created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WikiDoc"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/wiki/reorder:
    patch:
      summary: Reorder wiki docs
      operationId: reorderWikiDocs
      tags: [Wiki]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [slugs]
              properties:
                slugs:
                  type: array
                  items: { type: string }
      responses:
        "200":
          $ref: "#/components/responses/Ok"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/wiki/{slug}:
    parameters:
      - $ref: "#/components/parameters/slug"

    get:
      summary: Get a wiki doc by slug
      operationId: getWikiDoc
      tags: [Wiki]
      responses:
        "200":
          description: Wiki doc detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WikiDoc"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      summary: Create or update a wiki doc at a specific slug
      operationId: saveWikiDoc
      tags: [Wiki]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveWikiDocInput"
      responses:
        "200":
          description: Doc saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WikiDoc"
        "422":
          $ref: "#/components/responses/ValidationError"

    delete:
      summary: Delete a wiki doc
      operationId: deleteWikiDoc
      tags: [Wiki]
      responses:
        "200":
          description: Doc deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  slug: { type: string }
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Roadmap ───────────────────────────────────────────────────────────────
  /api/roadmap:
    get:
      summary: Get the full roadmap
      operationId: getRoadmap
      tags: [Roadmap]
      responses:
        "200":
          description: Full roadmap with cards, rows, and columns
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Roadmap"

    put:
      summary: Save the full roadmap (bulk update)
      operationId: saveRoadmap
      tags: [Roadmap]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Roadmap"
      responses:
        "200":
          description: Updated roadmap
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Roadmap"
  /api/roadmap/cards:
    post:
      summary: Add a roadmap card
      operationId: createRoadmapCard
      tags: [Roadmap]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCardInput"
      responses:
        "201":
          description: Card created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoadmapCard"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/roadmap/cards/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string }

    put:
      summary: Update a roadmap card
      operationId: updateRoadmapCard
      tags: [Roadmap]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCardInput"
      responses:
        "200":
          description: Updated card
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoadmapCard"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"

    delete:
      summary: Delete a roadmap card
      operationId: deleteRoadmapCard
      tags: [Roadmap]
      responses:
        "200":
          description: Card deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  id: { type: string }
        "404":
          $ref: "#/components/responses/NotFound"
  /api/roadmap/cards/{id}/move:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string }
    patch:
      summary: Move a roadmap card
      operationId: moveRoadmapCard
      tags: [Roadmap]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MoveCardInput"
      responses:
        "200":
          description: Moved card
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoadmapCard"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/roadmap/rows:
    post:
      summary: Add a roadmap row
      operationId: createRoadmapRow
      tags: [Roadmap]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRowInput"
      responses:
        "201":
          description: Row created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoadmapRow"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/roadmap/rows/reorder:
    patch:
      summary: Reorder roadmap rows
      operationId: reorderRoadmapRows
      tags: [Roadmap]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [labels]
              properties:
                labels:
                  type: array
                  items: { type: string }
      responses:
        "200":
          $ref: "#/components/responses/Ok"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/roadmap/rows/{label}:
    parameters:
      - name: label
        in: path
        required: true
        schema: { type: string }

    put:
      summary: Update a roadmap row
      operationId: updateRoadmapRow
      tags: [Roadmap]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRowInput"
      responses:
        "200":
          description: Updated row
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoadmapRow"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"

    delete:
      summary: Delete a roadmap row
      operationId: deleteRoadmapRow
      tags: [Roadmap]
      responses:
        "200":
          description: Row deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  label: { type: string }
        "404":
          $ref: "#/components/responses/NotFound"
  /api/roadmap/columns:
    post:
      summary: Add a roadmap column
      operationId: createRoadmapColumn
      tags: [Roadmap]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, minLength: 1 }
      responses:
        "201":
          description: Column created
          content:
            application/json:
              schema:
                type: object
                properties:
                  columns:
                    type: array
                    items: { type: string }
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/roadmap/columns/reorder:
    patch:
      summary: Reorder roadmap columns
      operationId: reorderRoadmapColumns
      tags: [Roadmap]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [names]
              properties:
                names:
                  type: array
                  items: { type: string }
      responses:
        "200":
          $ref: "#/components/responses/Ok"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/roadmap/columns/{name}:
    parameters:
      - name: name
        in: path
        required: true
        schema: { type: string }

    put:
      summary: Rename a roadmap column
      operationId: updateRoadmapColumn
      tags: [Roadmap]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, minLength: 1 }
      responses:
        "200":
          description: Updated columns
          content:
            application/json:
              schema:
                type: object
                properties:
                  columns:
                    type: array
                    items: { type: string }
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"

    delete:
      summary: Delete a roadmap column
      operationId: deleteRoadmapColumn
      tags: [Roadmap]
      responses:
        "200":
          description: Column deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  name: { type: string }
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Search ────────────────────────────────────────────────────────────────
...
components:
...
  parameters:
    page:
      name: page
      in: query
      schema: { type: integer, default: 1, minimum: 1 }
      description: Page number
...
    slug:
      name: slug
      in: path
      required: true
      schema: { type: string }
...
  responses:
    Ok:
      description: Success
      content:
        application/json:
          schema:
            type: object
            properties:
              ok: { type: boolean, example: true }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string }
...
    ValidationError:
      description: Validation error (422)
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string, description: "Semicolon-separated list of field-level errors" }
...
  schemas:
...
    WikiDoc:
      type: object
      properties:
        slug: { type: string }
        title: { type: string }
        description: { type: string, nullable: true }
        icon: { type: string, nullable: true }
        updated_at: { type: string, description: ISO 8601 date }
        body: { type: string }
        order: { type: number }
        dirty: { type: boolean, description: Whether the doc has unsaved changes }
    PaginatedWikiDocs:
      type: object
      properties:
        docs:
          type: array
          items: { $ref: "#/components/schemas/WikiDoc" }
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
        total_pages: { type: integer }
    SaveWikiDocInput:
      type: object
      required: [title]
      properties:
        title: { type: string, minLength: 1 }
        icon: { type: string, nullable: true }
        body: { type: string, default: "" }

    # ── Roadmap ────────────────────────────────────────────────────────────
    Roadmap:
      type: object
      properties:
        title: { type: string }
        columns:
          type: array
          items: { type: string }
        rows:
          type: array
          items: { $ref: "#/components/schemas/RoadmapRow" }
        cards:
          type: array
          items: { $ref: "#/components/schemas/RoadmapCard" }
        modified: { type: string, nullable: true }
    RoadmapCard:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        notes: { type: string }
        column: { type: string }
        row: { type: string }
        order: { type: number }
        feature_slug: { type: string, nullable: true }
    RoadmapRow:
      type: object
      properties:
        label: { type: string }
        color: { type: string, nullable: true }
    CreateCardInput:
      type: object
      required: [title]
      properties:
        title: { type: string, minLength: 1 }
        notes: { type: string }
        column: { type: string }
        row: { type: string }
        order: { type: number }
    UpdateCardInput:
      type: object
      properties:
        title: { type: string }
        notes: { type: string }
        column: { type: string }
        row: { type: string }
        order: { type: number }
        feature_slug: { type: string, nullable: true }
    MoveCardInput:
      type: object
      properties:
        column: { type: string }
        row: { type: string }
        order: { type: number }
    CreateRowInput:
      type: object
      required: [label]
      properties:
        label: { type: string, minLength: 1 }
        color: { type: string, nullable: true }
    UpdateRowInput:
      type: object
      properties:
        label: { type: string }
        color: { type: string, nullable: true }

    # ── Reorder ────────────────────────────────────────────────────────────
...